refactor: 컴포넌트 최상단 index.jsx > 컴포넌트명으로 수정
