+++ src/pages/SchoolPage/SchoolPage.jsx
