+++ src/routes/mainRoutes.jsx
@@ -2,13 +2,14 @@ import { Navigate } from 'react-router-dom';
 
 import MainLayout from '../layout/MainLayout';
 import { routes } from '../constants/routes';
-import MainPage from '../pages/MainPage';
-import SchoolPage from '../pages/SchoolPage';
-import MyPage from '../pages/MyPage';
+
 import AboutPage from '../pages/AboutPage/AboutPage';
 import PrivacyPage from '../pages/TermsPage/PrivacyPage';
 import TosPage from '../pages/TermsPage/TosPage';
 import RegionPage from '../pages/RegionPage/RegionPage';
+import SchoolPage from '@/pages/SchoolPage/SchoolPage';
+import MainPage from '@/pages/MainPage/MainPage';
+import MyPage from '@/pages/MyPage/MyPage';
 
 export const mainRoutes = {
 	path: '/',
